"ci delete" command was improved with new switches

+++ src/Configurator/Drill4Net.Configurator/src/Commands/CI/CiDeleteCommand.cs
@@ -20,28 +20,40 @@ namespace Drill4Net.Configurator
             if (_desc == null)
                 return Task.FromResult(false);
 
-            //delete the config
-            var cfgDir = _rep.GetCiDirectory();
-            var cfgPath = GetParameter(CoreConstants.ARGUMENT_CONFIG_PATH, false);
+            bool res = false;
+            var doNotDeleteIntegration = IsSwitchSet('I');
+            var deleteOnlyIntegration = IsSwitchSet('i');
 
-            bool res;
+            //delete the config
+            var cfgPath = GetParameter(CoreConstants.ARGUMENT_CONFIG_PATH, false);
             if (string.IsNullOrWhiteSpace(cfgPath))
             {
-                res = _cmdHelper.DeleteConfig<CiOptions>(CoreConstants.SUBSYSTEM_CI, cfgDir, _desc, out cfgPath);
+                var cfgDir = _rep.GetCiDirectory();
+                if (deleteOnlyIntegration)
+                {
+                    res = _cmdHelper.GetSourceConfigPath<CiOptions>(CoreConstants.SUBSYSTEM_CI, cfgDir, _desc, out cfgPath,
+                        out var _, out var error);
+                }
+                else
+                {
+                    res = _cmdHelper.DeleteConfig<CiOptions>(CoreConstants.SUBSYSTEM_CI, cfgDir, _desc, out cfgPath);
+                }
                 if (!res)
                     return Task.FromResult(false);
             }
 
             //delete the integrations with IDE
-            var ide = new IdeConfigurator(_rep);
-            var solutionDir = GetParameter(CoreConstants.ARGUMENT_SOURCE_DIR, false);
-            if (string.IsNullOrWhiteSpace(solutionDir))
+            if (!doNotDeleteIntegration)
             {
-                if (!_cli.AskDirectory("", out solutionDir, ide.GetDefaultProjectSourcesDirectory(), true))
-                    return Task.FromResult(true);
+                var ide = new IdeConfigurator(_rep);
+                var solutionDir = GetParameter(CoreConstants.ARGUMENT_SOURCE_DIR, false);
+                if (string.IsNullOrWhiteSpace(solutionDir))
+                {
+                    if (!_cli.AskDirectory("", out solutionDir, ide.GetDefaultProjectSourcesDirectory(), true))
+                        return Task.FromResult(true);
+                }
+                res = ide.DeleteInjections(solutionDir, cfgPath);
             }
-            res = ide.DeleteInjections(solutionDir, cfgPath);
-
             return Task.FromResult(res);
         }
